+++ setup.py
@@ -8,7 +8,7 @@ with open(path.join(this_directory, 'README.md'), encoding='utf-8') as f:
 
 setup(
     name="clara",
-    version="0.11.0", #Increse y, when gateway and archnetws are created. x.y.z
+    version="0.11.1", #Increse y, when gateway and archnetws are created. x.y.z
     author="Sergio Branco | The Architech",
     author_email="asergio.branco@gmail.com",
     description="",
